requirements.txt
deepspeed==0.9.5
einops==0.7.0
fastapi==0.104.1
flash_attn==2.2.3.post2
gradio
huggingface_hub==0.17.2
nltk==3.8.1
numpy==1.26.2
ogb==1.3.6
openai==0.28.0
packaging==23.2
peft==0.5.0
Pillow==10.1.0
pydantic==1.10.12
python_Levenshtein==0.23.0
rdkit==2023.3.3
Requests==2.31.0
rouge_score==0.1.2
scikit_learn==1.3.0
selfies==2.1.1
shortuuid==1.0.11
torch_geometric==2.3.1
tqdm==4.66.1
transformers==4.33.2
uvicorn==0.24.0.post1